Reorganizing test packages
Cleaned up imports
